Made steps more modular
