+++ package.json
@@ -30,6 +30,7 @@
     "moment": "^2.30.1",
     "node-fetch": "^2.7.0",
     "path-sort": "^0.1.0",
+    "punycode": "^2.3.1",
     "vinyl-source-stream": "^2.0.0",
     "yargs": "^17.7.2"
   },
